+++ tests/eval_files/async111.py
@@ -77,10 +77,10 @@ with trio.open_nursery() as nursery:
         async with trio.open_process() as bar_2:
             nursery.start(bar_2)  # safe
 
-# specifically check for *trio*.open_nursery
+# `import trio as noterror` -- open_nursery resolves to canonical qualname
 with noterror.open_nursery() as nursery:
     with trio.open("") as bar:
-        nursery.start(bar)
+        nursery.start(bar)  # error: 22, line-1, line-2, "bar", "start"
 
 # specifically check for trio.*open_nursery*
 with trio.open_nurse() as nursery: