+++ tests/test_config_and_args.py
@@ -491,9 +491,8 @@ def test_disable_noqa_ast(
     out, err = capsys.readouterr()
     assert not err
     assert (
-        out
-        == "./example.py:1:1: ASYNC106 trio must be imported with `import trio` for the"
-        " linter to work.\n"
+        out == "./example.py:1:1: ASYNC106 trio should be imported with `import trio`"
+        " for consistency.\n"
     )